+++ src/js/app/main/component/no-permission.vue
@@ -1,12 +1,3 @@
 <template>
-    <div class="hv-center text-muted f20">{{note}}</div>
+    <div class="hv-center text-muted f20">{{permissionNote}}</div>
 </template>
-<script>
-    export default{
-        computed: {
-            note(){
-                return this.$route.query.note;
-            }
-        }
-    };
-</script>